feat(lee): clean ancien fonctionnement
